HotelManolo/backend/requirements.txt
Flask
flask-cors
pymongo
pytesseract
pdf2image
Pillow
python-dotenv
